+++ TopsortAnalytics/src/main/java/com/topsort/analytics/Cache.kt
@@ -3,9 +3,9 @@ package com.topsort.analytics
 import android.content.Context
 import android.content.SharedPreferences
 import android.text.TextUtils
-import com.topsort.analytics.model.events.ClickEvent
-import com.topsort.analytics.model.events.ImpressionEvent
-import com.topsort.analytics.model.events.PurchaseEvent
+import com.topsort.analytics.model.ClickEvent
+import com.topsort.analytics.model.ImpressionEvent
+import com.topsort.analytics.model.PurchaseEvent
 import java.util.*
 
 private const val preferencesName = "TOPSORT_EVENTS_CACHE"
