+++ TopsortAnalytics/src/main/java/com/topsort/analytics/model/PurchaseEvent.kt
@@ -1,4 +1,4 @@
-package com.topsort.analytics.model.events
+package com.topsort.analytics.model
 
 import androidx.annotation.IntRange
 import com.topsort.analytics.core.getIntOrNull
