+++ TopsortAnalytics/src/test/java/com/topsort/analytics/TestObjects.kt
@@ -2,13 +2,13 @@ package com.topsort.analytics
 
 import com.topsort.analytics.core.eventNow
 import com.topsort.analytics.core.randomId
-import com.topsort.analytics.model.events.Click
-import com.topsort.analytics.model.events.Entity
-import com.topsort.analytics.model.events.EntityType
-import com.topsort.analytics.model.events.Impression
-import com.topsort.analytics.model.events.Placement
-import com.topsort.analytics.model.events.Purchase
-import com.topsort.analytics.model.events.PurchasedItem
+import com.topsort.analytics.model.Click
+import com.topsort.analytics.model.Entity
+import com.topsort.analytics.model.EntityType
+import com.topsort.analytics.model.Impression
+import com.topsort.analytics.model.Placement
+import com.topsort.analytics.model.Purchase
+import com.topsort.analytics.model.PurchasedItem
 
 fun getClickPromoted() : Click {
     return Click.Factory.buildPromoted(
