+++ app/src/main/java/com/topsort/analytics/JavaSampleActivity.java
@@ -5,10 +5,10 @@ import android.os.Bundle;
 import androidx.activity.ComponentActivity;
 import androidx.annotation.Nullable;
 
-import com.topsort.analytics.model.events.Entity;
-import com.topsort.analytics.model.events.EntityType;
-import com.topsort.analytics.model.events.Placement;
-import com.topsort.analytics.model.events.PurchasedItem;
+import com.topsort.analytics.model.Entity;
+import com.topsort.analytics.model.EntityType;
+import com.topsort.analytics.model.Placement;
+import com.topsort.analytics.model.PurchasedItem;
 
 import java.util.Collections;
